Refactor AuthJwtOdooEnv dependency to handle unauthenticated requests
Previously, the AuthJwtOdooEnv dependency did not handle unauthenticated
requests. This update adds support for unauthenticated requests by
checking if the request has authentication and allowing them when
'allow_unauthenticated' is set to True. In such cases, the environment
is returned with the base.public_user context.

Additionally, the code was refactored to simplify the logic and remove
an unnecessary assignment to env.uid.

This change benefits users by providing a more flexible authentication
mechanism and allowing unauthenticated requests when needed.

Fixes #123